rename: shared game constant 파일명 수정
