+++ src/entities/shared-game/model/shared-game.constant.ts
